fix(zrc-1): token_id bug
`token_id` count is bugged as it takes reference from `total_supply`. If `token_id = 1` is Burned while `total_supply = 2`. The `Mint` procedure will be broken and unusable. Solution is to add an additional field `token_id_count` that does not decrease, once a `token_id` is used, it will not be able to be used ever again.